feat :: error 예외처리
