+++ src/hooks/Auth/useSignup.ts
@@ -3,7 +3,7 @@ import { B1ndToast } from "@b1nd/b1nd-toastify";
 import { Signup, SignupAgree, SignUpModal } from "src/types/Signup/signup.type";
 import patternCheck from "src/utils/Check/patternCheck";
 // import * as Sentry from "sentry/react"
-import { useMemberSignUp } from "src/queries/Members/member.query";
+import { useMemberSignUp } from "src/queries/Member/member.query";
 import ErrorHandler from "src/utils/Error/ErrorHandler";
 import axios from "axios";
 import { useReqAuthCode,useSendAuthCode } from "src/queries/Member/member.query";
