+++ src/hooks/Bus/useApplyBus.ts
@@ -20,11 +20,7 @@ const useApplyBus = () => {
   const { 
     data: myBusData, 
     isLoading: myBusDataIsLoading,
-  } = useGetMyBusQuery({
-    suspense: true,
-    staleTime: 1000 * 30,
-    cacheTime: 1000 * 60,
-  });
+  } = useGetMyBusQuery();
 
   const postMyBusMutation = usePostMyBusMutation();
   const patchMyBusMutation = usePatchMyBusMutation();
