+++ src/hooks/Leave/useApplyLeave.ts
@@ -15,10 +15,7 @@ import dateTransform from "src/utils/Transform/dateTransform";
 const useApplyLeave = () => {
   const queryClient = useQueryClient();
 
-  const appliedLeaves = useGetMyLeavesQuery({
-    staleTime: 1000 * 30,
-    cacheTime: 1000 * 60,
-  }).data?.data;
+  const appliedLeaves = useGetMyLeavesQuery().data?.data;
 
   const postApplyLeaveMutation = usePostApplyLeaveMutation();
   const deleteApplyLeaveMutation = useDeleteApplyLeaveMutation();
