+++ src/queries/Banner/banner.query.ts
@@ -1,8 +1,9 @@
 import { AxiosError } from "axios";
 import { useQuery, UseQueryOptions, UseQueryResult } from "react-query";
-import bannerApi from "src/repositories/Banner/banner.repository";
+import bannerRepository from "src/repositories/Banner/banner.repository";
 import { BannersResponse } from "src/types/Banner/banner.type";
 import { QUERY_KEYS } from "../queryKey";
+import { B1ndToast } from "@b1nd/b1nd-toastify";
 
 export const useGetBannersQuery = (
   options?: UseQueryOptions<
@@ -12,4 +13,19 @@ export const useGetBannersQuery = (
     string
   >
 ): UseQueryResult<BannersResponse, AxiosError> =>
-  useQuery(QUERY_KEYS.banner.get, () => bannerApi.getBanners(), options);
+  useQuery(QUERY_KEYS.banner.get, 
+    () => bannerRepository.getBanners(),
+ {
+        
+        cacheTime: 1000 * 60 * 60,
+        staleTime: 1000 * 60 * 30,
+          ...options,
+           onError: (error:AxiosError) => {
+                  if(error.status == 500){
+                    B1ndToast.showError("서버 에러발생");
+                    return;
+                  }
+                  window.location.reload();
+                },
+        }
+);
