+++ src/queries/Schedule/schedule.query.ts
@@ -3,6 +3,8 @@ import { useQuery, UseQueryOptions, UseQueryResult } from "react-query";
 import scheduleRepository from "src/repositories/Schedule/schedule.repository";
 import { ScheduleResponse } from "src/types/Schedule/schedule.type";
 import { QUERY_KEYS } from "../queryKey";
+import { B1ndToast } from "@b1nd/b1nd-toastify";
+
 
 export const useGetScheduleQuery = (
   options?: UseQueryOptions<
@@ -20,6 +22,12 @@ export const useGetScheduleQuery = (
       staleTime: 1000 * 60 * 5,
       cacheTime: 1000 * 60 * 10,
       ...options,
+      onError: (error:AxiosError) => {
+        if(error.status == 500){
+          B1ndToast.showError("서버 에러발생");
+          return;
+        }
+        window.location.reload();
+      },
     }
-    
   );
